fastAPI/requirements.txt
pandas==2.3.2
fastapi==0.116.1
fredapi==0.5.2
httpx==0.28.1
numpy==2.3.2
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
scipy==1.16.1
SQLAlchemy==2.0.23
torch==2.8.0
yfinance==0.2.65
transformers==4.55.2
uvicorn
psycopg2